extract media/pacer package, finalize ts pacer
* Extract pacer logic, discard, stats, and logging into a new media/pacer package shared by RTP and MPEG-TS pacers.
* Add PcrUnwrapper/PcrGapDetector to media/mpegts
* Add a clock helper to media/rtp
* Add unit tests for all mpegts timing utilities (PcrToDuration, DurationToPcr, PtsToDuration, PcrDiff, ExtractPCR, ExtractPTS).
* Change DurationToPcr to return uint64 and PcrGapDetector.Threshold to uint64 for consistency with PCR semantics.
* Rename CHD (channel delay) to JBD (jitter buffer delay) in pacer stats for clarity and consistency.
* Add InitDefaults() to config structs
* Tune defaults

+++ media/mpegts/ts_util.go
@@ -93,8 +93,8 @@ func PcrToDuration(diff uint64) time.Duration {
 
 // DurationToPcr converts a time.Duration to PCR ticks (27 MHz clock).
 // Inverse of PcrToDuration: t = d * 27 / µs.
-func DurationToPcr(d time.Duration) int64 {
-	return int64(d) * 27 / int64(time.Microsecond)
+func DurationToPcr(d time.Duration) uint64 {
+	return uint64(int64(d) * 27 / int64(time.Microsecond))
 }
 
 func PtsToDuration(diff uint64) time.Duration {
